Add ignore option to py3k linter.
This makes it easier to deal with from future import absolute_import
errors if you are already in py27.
